requirements.txt
# websockets
# SimpleWebSocketServer
python-socketio
eventlet
peewee
mysqlclient